Remove mergify from repository
While it initially seemed handy, the danger of automatically merging a
broken PR is too great in this specific project.  This is because the
cirrus-ci_retrospective action must necessarily execute after cirrus-ci
is finished.  Due to limitations of the github issue-labling
system/permissions, it is likely mergify could merge before the
self-test run of cirrus-ci_retrospective on a PR has completed.

Signed-off-by: Chris Evich <cevich@redhat.com>