Trigger task on action success
Previously, if the cirrus-ci_retrospective action and integration test
failed, it would not block accidental merging of a PR.  This commit
adds a cirrus-ci task that does not start automatically upon push.
Instead, it will be state-validated and executed *only* after
successfull execution of the integration test.

Also add a small github action helper library, unit-tests, and
verify they pass for PRs and releases.  These all had to be
manually verified in a 'sandbox' repo., as they will not
execute in a PR (until it's merged).

Signed-off-by: Chris Evich <cevich@redhat.com>