Avoid build status race
It's possible that Cirrus-CI could re-execute a build/task after
previously having completed.  For example, due to a manual trigger or
re-run of a task.  This can result in the workflow triggered by the
first completion event, finding a build not in a final state.  If this
condition is detected, the workflow should not take any further action.
Presumably the in-flight (manual or re-run) will complete again at some
future time, causing the `on: check_suite: completed' workflow to execute
again.

Signed-off-by: Chris Evich <cevich@redhat.com>