Implement smoke-test for cirrus-ci_retrospective
Previously, if breaking changes were made to the cirrus-ci_retrospective
container image (or it's contents), failures wouldn't appear until after
the offending PR was merged.  Fix this, by detecting a running PR (using
formerly successful container image).  Build a test container from the
PR's code and use it to verify functionality and action outputs.  On
failure, the tracking (auto-updated) comment in the PR will be updated
with links and debugging info.

Signed-off-by: Chris Evich <cevich@redhat.com>