Implement key storage in kernel
Originally intended to do this, but saving the key directly into a file
was easier for an initial implementation (despite what the comment
said).  This commit actually implements storing the key in the kernel,
when the kernel supports keyring.  Otherwise the old/simple obfuscation
method is used.

Also fixed some spelling mistakes and improved a few comments

Signed-off-by: Chris Evich <cevich@redhat.com>