Sort nodes by "size" instead of alphabetical
...where "size" is defined as the sum of input and output
edges plus the number of matrix jobs. The effect is to
put the "big" boxes (lots of arrows, lots of jobs) down
at the bottom, and the "small" ones (ellipses) at top.
Makes for a somewhat more readable graph IMO.

(OBTW the key insight here is that 'dot' draws the nodes
top to bottom in input order.)

Signed-off-by: Ed Santiago <santiago@redhat.com>