Update CI Container packages + export env. vars.
As of the release of Fedora 33, the FindBin perl module is now provided
by a dedicated package.  Previously it was part of the perl-interpreter
package.

For some unit-tests, it may become necessary in the future to utilize
Cirrus-CI env. vars. if they are set.  Ensure the top-level runner
exports them when they are available.

Signed-off-by: Chris Evich <cevich@redhat.com>