Fix code-context references
Previously due to practical reasons, the unit-tests did a very poor
jobs checking the code-context information shown by functions like
`die()`.  Prior to this commit, the stack-level used was too great,
resulting in unhelpful output such as:

`ERROR: blah blah blah (<stdin>:0 in ())`

Fix the stack-level and context output.  Add a helper-script
and new unit-test to confirm it does not break again in the future.

Signed-off-by: Chris Evich <cevich@redhat.com>