Simplify global automation env. vars
tl;dr; Helps simplify this nonsense:
https://github.com/containers/podman/pull/8046/files#diff-87b53ab90b08478b1624b8f218f6d0b6eab834f2cf81dd73581540c3a0dfb63cR9-R14

The automation libraries may be located in different places
based on install-time options.  In order for a callers to utilize the
libraries, they must first load the definition of `$AUTOMATION_LIB_PATH`.

Prior to this commit, this env. var definition could be in one/more
locations depending upon the platform.  This arrangement added
unnecessary complexity. Fix this with a **non-backward compatible
change** such that the only file that needs to be loaded for all
platforms, is in a well defined location:

`/etc/automation_environment`

If needed, this filename may be changed at install-time by overriding
`$INSTALL_ENV_FILEPATH`

Signed-off-by: Chris Evich <cevich@redhat.com>