Skopeo/Buildah/Podman multi-arch image GH actions
The `repository_details` action helps look up the proper
secret env. vars and repository namespace details based
on the workflow's matrix-item inputs.  In other words
it allows the build & push action (below) to push to multiple
image repositories without being concerned with the login
or namespace details.

The `multi-arch-build-push` action builds a Skopeo, Buildah,
or Podman image for multiple architectures.  If successful,
it is pushed as 'latest' to a repository namespace.  It then
obtains the skopeo|buildah|podman version number, and conditionally
pushes if a version-tagged image does not already exist.

Signed-off-by: Chris Evich <cevich@redhat.com>