Remove unmaintained/disused ephemeral_gpg tool
This tool was handy/useful for securely containing a gpg environment.
It was originally intended to assist in automated releases of podman,
but that effort was abandoned due to implementation time and difficulty.
Since the tool is security related and unmaintained, it's safer to
remove remove it to avoid giving anyone the wrong impression of its
status.

Signed-off-by: Chris Evich <cevich@redhat.com>