Remove unmaintained/disused cirrus-ci_asr tool
This tool was handy/useful for monitoring the status of a running build
from the command-line.  However, to my knowledge it's not actually been
used in quite a while.  Recently dependabot alerted on a security update
for the websockets python module, but fixing it broke CI.  Simply
remove the script instead.

Signed-off-by: Chris Evich <cevich@redhat.com>