Fix $ARCHES export + special-chars in $BUILD_ARGS
It's necessary for downstream usage that both a --modcmd and --prepcmd
script have access to the parsed value from --arches.  Prior to this
commit, the value was stored in an array.  Unf., many versions of bash
do not yet support export of array variables.  This commit converts it
into a simple space-separated string and adds a test to confirm.

While fixing this, it was discovered that the handling of \$BUILD_ARGS
was also wrong.  Both quoting and any use of embedded special characters
(like whitespace) would not be preserved.  It was also marked as a
variable exported to to --prep/--modcmd, which is not possible.  Fix
this by removing it from the export list, and tweak the processing and
use of \$BUILD_ARGS in general.  Add a test to confirm proper handling
of both quotes and special characters.

Lastly, some seemingly unrelated changes were made to simplify the
addition of the above tests.

Signed-off-by: Chris Evich <cevich@redhat.com>