Fix automatic inclusion of local arch
When executing build-push.sh with the `--arches=*` option, the script
automatically includes the local architecture by default.  This may
be counter-intuitive and is contrary to the documentation for this
option.  Worse, if the local architecture is specified as an
`--arches` argument, buildah will build TWO images for it. Fix this
by excluding the default local-arch value when processing the
`--arches` command-line option.

Signed-off-by: Chris Evich <cevich@redhat.com>