GHA: Shadow common-lib die/warn/dbg functions
Github-action workflow runs can consume additional output "sugar" values
and use them to annotate output in the UI.  When using the github
library, shadow the common-lib functions to add in this extra metadata.
Also, fix use of the debug calls so they are useful despite
ACTIONS_STEP_DEBUG being set or not.

Signed-off-by: Chris Evich <cevich@redhat.com>