CI VM OS updates as 'major' vs 'compatibility'
Prior to this commit, a major OS update (e.g. F37 -> F38) was flagged to
renovate as an 'incompatible' change to propose.  This was done to allow
additional manual testing and scrutiny.  However, now that CI VM update
PRs are directly assigned to me, I can directly keep a close eye on
them.  Including when the OS name strings need to be updated.  Further,
by adjusting the major/minor/patch labeling of the various image
"version" components, they will better represent the size of the changes.
This is reflected in the titles of the PRs opened by renovate.

Signed-off-by: Chris Evich <cevich@redhat.com>