fix broken renovate config
checked with:
podman run -it \
-v ./renovate/defaults.json5:/usr/src/app/renovate.json5:z \
docker.io/renovate/renovate:latest \
renovate-config-validator

Due the nested packageRules section in golang the auto migration is not
working correctly and caused an error for us. This caused renovate to
propse PRs without the proper settings.
Fix the config by (hopefully) migrating correctly to the new format.
The nested packageRule is now on the same level which should fix the
breakage. The config validator is happy now but I have no way of
actually testing if this still works correctly, I guess we will find
out.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>