Improve build-push error handling
Around the time of this commit, the automated multiarch manifest-list
builds for both skopeo and podman have been failing somewhere in the
`build-push.sh` script.  The actual build appears to work fine, the
`tag-version.sh` mod-command runs fine, but the tag-search in
`get_manifest_tags()` (called by `push_images()`) fails with the error:

`jq: error (at <stdin>:29): Cannot iterate over null (null)`

Unfortunately the problem does not reproduce for me locally, nor can it
be reproduced using a dry-run build (`--nopush` bypasses the tag search.)
Improve debugging of this situation by moving the `if ((PUSH))` check and
adding an exception clause to display the would-be pushed images (and
tags).

Also:

* Simplify the `get_manifest_tags()` tag search by adjusting the jq filter
  to gracefully ignore an empty set of images and/or images without
  any list of names.  Rely on `push_images()` catching the empty-list
  and throwing an error.
* Add a comment regarding the need for the `confirm_arches` call
  after the `parallel_build` call in the main part of the script.
* Improve the debug-ability of `confirm_arches()` in the case of
  a bad/incomplete/unreadable manifest-list (see item above).
  Detect both inspect command errors and jq/pipeline errors.  In
  the case of jq/pipeline failure, show the input JSON to aid
  debugging.
* Improve variable-name consistency by removing many `_` prefixes.

Signed-off-by: Chris Evich <cevich@redhat.com>