Workaround Renovate PR force-push spam
Ref:
https://github.com/containers/buildah/pull/5007#issuecomment-1819695794

Attempt to limit excessive Renovate force-push activity related to
golang update PRs.  This is unlikely a perfect fix, and it may not
address the problem at all.  I believe there's a renovate and/or
configuration bug here, but lack the resources/time to investigate more
deeply - this likely requires a temp. reproduction repo. + mock golang
package.

Signed-off-by: Chris Evich <cevich@redhat.com>