Create, mount, and use local storage
Podman machine testing is very much storage-bound in terms of
performance.  The stock AWS setup uses networked storage for the system,
and a small local disk for `/tmp`.  However there is plenty of empty
space available on the local disk, and it's *MUCH* faster than network
storage.  Use this disk as the worker-user's home directory (where tests
run from).

Signed-off-by: Chris Evich <cevich@redhat.com>