Fix unmanaged crashed/inaccessible worker
If a worker instances is inaccessible for an extended period of time,
it's a sign it may have crashed or been compromised in some way.
Previously, due to the order of status checks this condition would not
be noticed for multiple days.  Fix this by relocating the `PW_MAX_HOURS`
check to the beginning of the worker-loop.  This will force-terminate
any timed-out instances regardless of all other status checks.

Signed-off-by: Chris Evich <cevich@redhat.com>