Fix possible max-tasks PW pool cascade failure
For integrity and safety reasons, there are multiple guardrails in place
to limit the potential damage of a rogue/broken/misconfigured worker
instance may cause. One of these restrictions is a maximum limit on the
number of tasks that a worker may execute. However, if the pool is
experiencing extraordinary utilization, it's possible that a large number
of workers could encounter this limit at/near the same time. Assuming the
pool load remains high, this will then further shorten the lifetime of
the remaining online instances.

Also:

* Double the limit on allowed tasks (12 was too small based on heavy
  utilization).
* Double the allowed setup time to account for network slowdowns.
* Show both the soft and hard uptime limits for each worker.
* Issue warning if worker exceeds soft uptime limit.

Signed-off-by: Chris Evich <cevich@redhat.com>