renovate: update to go 1.22
We have pinned the renovate go version to the lowest version we support
as otherwise it will create PRs that update to new go version which we
always want to take care of manually as it included more changes
usually. While it doesn't prevent renovate from creating these PRs they
always fail as it cannot update to a newer go version so it is clear to
reviewers what is going on.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>